requirements.txt
# Requirements per il progetto

requests
numpy
beautifulsoup4
googlesearch-python
openai
python-dotenv
langchain_openai
langchain
langchain-community
streamlit
